Generalized config.json and rename app.js to index.js
