+++ index.js
@@ -1,11 +1,20 @@
-'use strict';
+const express = require('express');
+const bodyParser = require('body-parser');
 
-/**
- * Adds commas to a number
- * @param {number} number
- * @param {string} locale
- * @return {string}
- */
-module.exports = function(number, locale) {
-    return number.toLocaleString(locale);
-};
+const config = require('./config.json');
+const workingHours = require('./routes/workinghour');
+const appointments = require('./routes/appointment');
+
+const app = express();
+
+
+app.use(bodyParser.json());
+app.use(bodyParser.urlencoded({ extended: true }));
+
+app.use('/workingHours', workingHours);
+app.use('/appointments', appointments);
+
+app.listen(config.PORT, "0.0.0.0");
+console.log('we live on ' + config.PORT);
+
+module.exports = app;
