+++ package.json
@@ -1,10 +1,10 @@
 {
-  "name": "luma-eng-interview",
+  "name": "luma-eng-interview-xingcheng-sun",
   "version": "0.1.0",
   "description": "A REST API that enables a simple scheduling system that manages doctor availabilities and allows patients to book appointments.",
   "main": "index.js",
   "scripts": {
-    "start": "nodemon app.js",
+    "start": "nodemon index.js",
     "test": "mocha --reporter spec"
   },
   "repository": {
